pyproject.toml
# NOTES:
# - Currently using Python 3.10 for development
# - Currently, the devices in fullbindings are the ones that work
# - Fix PyQt6/PySide6 xcb plugin error on Linux by installing:
#  sudo apt install libxcb-cursor0

[build-system]
requires = ["setuptools>=61", "setuptools_scm"]
build-backend = "setuptools.build_meta"

[project]
name = "epicpy"
version = "2026.3.5.2"
authors = [
    { name = "Travis L. Seymour, PhD", email = "nogard@ucsc.edu" }
]
description = "Python interface to the EPIC computational cognitive architecture simulation environment."
readme = "ReadMe.md"
requires-python = ">=3.10,<3.14"
license = { text = "GNU General Public License v3 or later (GPLv3+)" }
classifiers = [
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Operating System :: OS Independent",
    "Natural Language :: English",
    "Topic :: Education",
    "Topic :: Scientific/Engineering",
    ]

dependencies = [
    "loguru>=0.6.0,<0.8",
    "qtpy>=2.3,<3.0",
    "PySide6-Essentials>=6.8.3",
    "PySide6-Addons>=6.8.3",
    "python-dateutil>=2.8.2,<3.0",
    "ulid>=1.1,<2.0",
    "ulid2>=0.3.0,<1.0",
    "numpy>=1.24,<3.0",
    "scipy>=1.9,<2.0",
    "pingouin>=0.5.2,<1.0",
    "plum-dispatch>=2.5.0,<=2.6.0; sys_platform != 'win32'",
    "plum-dispatch>=2.5.0,<=2.6.0; sys_platform == 'win32'",
    "ascii_frame>=0.0.1,<1.0",
    "pandas>=2.0,<3.0",
    "polars>=1.32.3,<2.0",
    "pyqtdarktheme", # ">=2.1.0,<3.0",
    "pyqtgraph>=0.13.3,<1.0",
    "qdarkstyle>=3.1.0,<4.0",
    "platformdirs>=3.0,<4.0",
    "requests>=2.28,<3.0",
    "fastparquet>=2024.11.0",
    "pyarrow>=21.0.0",
    "pywin32; sys_platform == 'win32'",
    "rich>=13.4,<14",
    "networkx>=3.1,<4",
    "graphviz>=0.20.3,<1",
    'tomli>=2.0.1,<3; python_version < "3.11"',
    "epiclibcpp>=0.1.4,<1",
    "munch>=4.0.0,<5",
    "epicpydevicelib>=0.1.0",
    "psutil>=5.9,<7.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.13.0",
    "plumbum>=1.8.1",
    "pytest>=7.2.2",
    "qt6-tools==6.4.3.1.3"
]

[project.urls]
"Homepage" = "https://github.com/travisseymour/EPICpy"

[tool.setuptools.packages.find]
exclude = [
    "contrib", "docs", "tests*", "test*", "build", "epicpy.egg-info"
]

[project.scripts]
EPICpy = "epicpy.main:main"
epicpy = "epicpy.main:main"

[tool.setuptools]
include-package-data = true
zip-safe = false

[tool.ruff]
line-length = 130
exclude = ["epicpy/uifiles"]

[tool.ruff.lint]
select = [
    "E",
    "F",
    "I",
]
ignore = [
    "E402",
    "E731"
]